pyproject.toml
[tool.poetry]
name = "mtb"
version = "0.1.0"
description = ""
authors = ["Mika Qvist <mika.k.qvist@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
hypothesis = "^6.68.2"
nicegui = "^1.2.2"
click = "^8.1.3"
svg-py = "^1.4.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mtb = 'mtb.cli:run'
